Get things working
This first pass ensures that:

    a) most if not all use-cases are covered
    b) things work

Currently, the system has some limitations such as hardcoded input file,
hardcoded accuracy threshold, coordinates are 0-based (which may be
weird and look like it's off-by-one) and the order in which the detected
invaders are output to the screen is determined by dict order in Python.

A second pass will implement a proper CLI, with options for threshold,
input file, order (accuracy or position), clean up some code (like
radar-factory interactions) and ensure the test input is tested in the
test suite.